pyproject.toml
[tool.poetry]
name = "tapestry"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "tapestry", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
chromadb = "^0.6.0"
flask = "^3.1.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
pillow = "^11.0.0"
torchserve = "0.8.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
timm = "^1.0.12"
open-clip-torch = "^2.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
